% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locgff.R
\name{ExtractMd5}
\alias{ExtractMd5}
\title{Extract md5sum}
\usage{
ExtractMd5(mdfilePath, pattern)
}
\arguments{
\item{pattern}{A \code{character string} whether "gff" or "feature_table"}

\item{mdfile}{The md5sum file path.}
}
\value{
A \code{Character} md5sum string.
}
\description{
Extract md5sum for a given patterned files
}
\details{
md5sum for the "gff" and "feature_table" files
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
