% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenomicGene.R
\name{getGenomicGenes}
\alias{getGenomicGenes}
\alias{getGenesfGenomes}
\alias{getGenesfGeneids}
\title{Retrieve genomic genes}
\usage{
getGenomicGenes(KEGGSpe)

getGenesfGenomes(KEGGSpe)

getGenesfGeneids(KEGGSpe)
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}
}
\value{
A list. Each element represents a genome, and the gene IDs are in KEGG format.

A list. Each element represents a genome, and the gene IDs are in KEGG format.

A list. Each element represents a genome, and the gene IDs are in KEGG format.
}
\description{
Retrieve genomic genes.
}
\details{
Retrieve the whole genomic gene from a prokaryotic species. The function returns gene location in all genomes, including the cellular genoms and plasmids. For some gene, multiple location may return. If the genome has no gene featurs, "NULL" will be returned.
getGenesfGenomes(): retrieve whole genomic genes from NCBI genomes.
getGenesfGeneids(): retrieve whole genomic genes from NCBI gene IDs.
getGenomicGenes(): a wrapped of these top two function, which trys getGenesfGeneids() at first and then getGenesfGenomes() if the genome is from GenBank.
}
\examples{
## no genomic genes
noGenes <- getGenomicGenes('csu')
\dontrun{
## from GenBank
## one genome with three plasmids
hxaGenes <- getGenomicGenes('hxa')

## from RefSeq
## two genomes with two plasmids
draGenes <- getGenomicGenes('dra')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
