% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenomicInfo.R
\name{getFTPUrl}
\alias{getFTPUrl}
\title{Retrieve genome FTP URL}
\usage{
getFTPUrl(KEGGSpe, database = "GenBank")
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}

\item{database}{"GenBank" or "RefSeq".}
}
\value{
FTP url.
}
\description{
Get genome FTP url
}
\details{
Get the FTP url storing genome information from GenBank or RefSeq database. This function is mainly used to get the species genome assembly information.
}
\examples{
ecoliUrl <- getFTPUrl('eco')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

