% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenomicGene.R
\name{getGenomicGenes}
\alias{getGenomicGenes}
\title{Retrieve genomic genes}
\usage{
getGenomicGenes(KEGGSpe)
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}
}
\value{
A list. Each element represents a genome, and the gene IDs are in KEGG format.
}
\description{
Download genomic genes.
}
\details{
Download the whole genomic gene from a prokaryotic species. The function returns gene location in all genomes, including the cellular genoms and plasmids.
}
\examples{
\dontrun{
## from GenBank
## one genome with four plasmids
hxaGenes <- getGenomicGenes('hxa')

## from RefSeq
## two genomes with two plasmids
draGenes <- getGenomicGenes('dra')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

