% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locgff.R
\name{GetLocsfgff}
\alias{ExtractLocs}
\alias{GetLocsTag}
\alias{GetLocsfKEGGSpe}
\alias{GetLocsfgff}
\alias{download.Spegff}
\title{Get genomic gene locations from the gff file}
\usage{
GetLocsfgff(gffRawMat, genePrefix = character(0))

GetLocsfKEGGSpe(KEGGSpe)

GetLocsTag(annoStr)

ExtractLocs(gffRawMat)

download.Spegff(KEGGSpe, saveFolder)
}
\arguments{
\item{gffRawMat}{raw gff matrix.}

\item{genePrefix}{prefix to locus gene names.}

\item{KEGGSpe}{A KEGG species ID}

\item{annoStr}{character strings of gff annotation, which is sperated by ';'.}

\item{saveFolder}{A folder to save gff and md5sum check files. If the folder does not exist, then creat a one at first.}
}
\value{
GetLocsfgff(): a list of genomes containing gene location information. The locus_tags is used for the gene names.

GetLocsfKEGGSpe():  a list of genomes containing gene location information. The locus_tags is used for the gene names.

locus_tags

ExtractLocs(): 4-column matrix, 1st is the locus_tags, 2ed is the from locus, 3rd is the to locus, 4th is the strand.

download.Spegff(): download the gff and md5sum files.
}
\description{
Get genomic genes location
}
\details{
GetLocsfgff(): get genomic gene location information from gff matrix. The whole genomic gene are divided by chromosomes and plasmids (if the organism has).

GetLocsfKEGGSpe(): get genes locus of KEGG species from NCBI gff files. It trys the RefSeq database at first; if RefSeq is not found, then changes to the database to GenBank. The prefix of locus name is the abbreviation of KEGG genomes.

ExtractLocs(): extract gene location from gff raw file.

download.Spegff(): download gff and md5sum check files. If the md5sum check fails, download the files again until it passes.
}
\examples{
## read in the dra (Deinococcus radiodurans R1) gff gz file in local disk
gzPath <- system.file("extdata", "dra.gff.gz", package = "ProGenome")
dragff <- read.gff(gzPath, isurl = FALSE, isgz = TRUE)

## only extract whole locus without genome division
locsRawMat <- ExtractLocs(dragff)

## whole locus divided by genomes and plasmids
locusList <- GetLocsfgff(dragff, genePrefix = 'dra:')

## get dra genomic locus through FTP URL
draLocs <- GetLocsfKEGGSpe('dra')
\dontrun{
hxaLocs <- GetLocsfKEGGSpe('hxa')
csuLocs <- GetLocsfKEGGSpe('csu')}
\dontrun{
download.Spegff('eco', 'tmpEco')
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
read.gff read in gff files
}
\keyword{internal}

