% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pttrnt.R
\name{ExtractPtt}
\alias{ExtractPtt}
\alias{ExtractRnt}
\alias{write.ptt}
\alias{write.rnt}
\title{ptt and rnt}
\usage{
ExtractPtt(featureTable)

ExtractRnt(featureTable)

write.ptt(ptt, file)

write.rnt(rnt, file)
}
\arguments{
\item{featureTable}{A \code{matrix} represents the feature table retrieved from the NCBI ftp.}

\item{ptt}{A \code{matrix} indicating the ptt matrix.}

\item{file}{A \code{character vector} indicating the file path.}

\item{rnt}{A \code{matrix} indicating the rnt matrix}
}
\value{
\code{ExtractPtt()}: Extract ptt from the featureTable.

\code{ExtractRnt()}: Extract rnt from the featureTable.
}
\description{
Extract and write old ptt and rnt
}
\details{
\code{ExtractPtt()}: Extract ptt from the featureTable.

\code{ExtractRnt()}: Extract rnt from the featureTable.

\code{write.ptt()}: Write the ptt file.

\code{write.rnt()}: Write the rnt file.
}
\examples{
gzPath <- system.file('extdata', 'eco.feature_table.txt.gz', package = 'ProGenome')
ft <- read.gff(gzPath)

ptt <- ExtractPtt(ft)
rnt <- ExtractRnt(ft)

\dontrun{
write.ptt(ptt, 'eco.ptt')
write.rnt(rnt, 'eco.rnt')
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
