% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenomicInfo.R
\name{getSpeFtpUrl}
\alias{getSpeFtpUrl}
\alias{listFileFtpUrl}
\title{Retrieve genome FTP URL}
\usage{
getSpeFtpUrl(KEGGSpe, database = "GenBank")

listFileFtpUrl(ftpUrl)
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}

\item{database}{"GenBank" or "RefSeq".}

\item{ftpUrl}{FTP url that can be the output of function  getFtpUrl()}
}
\value{
FTP url.

A list of download FTP urls
}
\description{
Get genome FTP url
}
\details{
getFtpUrl(): get the FTP url storing genome information from GenBank or RefSeq database. This function is mainly used to get the species genome assembly information.
listFtpFileUrl(): list the file download FTP urls.
}
\examples{
ecoliUrl <- getSpeFtpUrl('eco')
ecoliFiles <- listFileFtpUrl(ecoliUrl)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}

