% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGenomicInfo.R
\name{getSpeFtpUrl}
\alias{getLocsfgff}
\alias{getSpeFtpUrl}
\alias{listFileFtpUrl}
\alias{read.gffUrl}
\title{Retrieve genome FTP URL}
\usage{
getSpeFtpUrl(KEGGSpe, database = "GenBank")

listFileFtpUrl(ftpUrl)

read.gffUrl(gffUrl)

getLocsfgff(KEGGSpe)
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}

\item{database}{"GenBank", "RefSeq", or c("GenBank", "RefSeq").}

\item{ftpUrl}{FTP url that can be the output of function  getFtpUrl()}

\item{gffUrl}{FTP url of one gff file.}
}
\value{
FTP url.

A list of download FTP urls

A table of raw gff file

A list of genomes containing gene location information. The locus_tags is used for the gene names.
}
\description{
Get genomic information NCBI FTP url
}
\details{
getFtpUrl(): get the FTP url storing genome information from GenBank or RefSeq database. This function is mainly used to get the species genome assembly information.
listFtpFileUrl(): list the file download FTP urls.
read.gffUrl(): read in raw gff file from a FTP url.
getLocsfgff(): get genomic gene location information from gff file. It trys the RefSeq database at first; if RefSeq is not found, then changes to the database to GenBank.
}
\examples{
## E. coli genome GenBank FTP
ecoliUrl <- getSpeFtpUrl('eco', c('GenBank', 'RefSeq'))
## list E. coli FTP files
ecoliFiles <- listFileFtpUrl(ecoliUrl[1])
## read in gff file
gffUrl <- ecoliFiles[grepl('gff', ecoliFiles)]
ecoligff <- read.gffUrl(gffUrl)
\dontrun{
draliUrl <- getSpeFtpUrl('dra')
draliFiles <- listFileFtpUrl(draliUrl)
}
draLocs <- getLocsfgff('dra')
\dontrun{
hxaLocs <- getLocsfgff('hxa')
csuLocs <- getLocsfgff('csu')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\references{
How to read in gz files into R from FTP URL \url{https://stackoverflow.com/questions/9548630/read-gzipped-csv-directly-from-a-url-in-r}
}

