% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locgff.R
\name{GetLocsfgff}
\alias{GetLocsfgff}
\title{Get genomic gene locations from the gff file}
\usage{
GetLocsfgff(KEGGSpe)
}
\arguments{
\item{KEGGSpe}{A KEGG species ID}
}
\value{
GetLocsfgff(): a list of genomes containing gene location information. The locus_tags is used for the gene names.
}
\description{
Get genomic genes location of KEGG species from NCBI gff files
}
\details{
getLocsfgff(): get genomic gene location information from gff file. It trys the RefSeq database at first; if RefSeq is not found, then changes to the database to GenBank.
}
\examples{
draLocs <- GetLocsfgff('dra')
\dontrun{
hxaLocs <- GetLocsfgff('hxa')
csuLocs <- GetLocsfgff('csu')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

