% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locgff.R
\name{GetLocsfgff}
\alias{GetLocsfgff}
\alias{GetLocsfKEGGSpe}
\alias{GetLocsTag}
\alias{ExtractLocs}
\alias{download.SpeAnno}
\title{Get genomic gene locations from the gff file}
\usage{
GetLocsfgff(gffRawMat, genePrefix = character(0))

GetLocsfKEGGSpe(KEGGSpe)

GetLocsTag(annoStr)

ExtractLocs(gffRawMat)

download.SpeAnno(KEGGSpe, pattern, saveFolder)
}
\arguments{
\item{gffRawMat}{raw gff matrix.}

\item{genePrefix}{prefix to locus gene names.}

\item{KEGGSpe}{A KEGG species ID}

\item{annoStr}{character strings of gff annotation, which is sperated by ';'.}

\item{pattern}{A \code{character string} whether "gff" or "feature_table"}

\item{saveFolder}{A folder to save gff and md5sum check files. If the folder does not exist, then creat a one at first.}
}
\value{
GetLocsfgff(): a list of genomes containing gene location information. The locus_tags is used for the gene names.

GetLocsfKEGGSpe():  a list of genomes containing gene location information. The locus_tags is used for the gene names.

a matrix. Locus_tags, old_locus_tags will also be return if provided.

ExtractLocs(): 4 or 5-column matrix, 1st (or 1st and 2ed) is the locus_tags, the last three are start position, end postion, and DNA strand.

download.SpeAnno(): download the gff, feature_table, or md5sum files.
}
\description{
Get genomic genes location
}
\details{
GetLocsfgff(): get genomic gene location information from gff matrix. The whole genomic gene are divided by chromosomes and plasmids (if the organism has).

GetLocsfKEGGSpe(): get genes locus of KEGG species from NCBI gff files. It trys the RefSeq database at first; if RefSeq is not found, then changes to the database to GenBank. The prefix of locus name is the abbreviation of KEGG genomes.

ExtractLocs(): extract gene location from gff raw file.

download.SpeAnno(): download gff/feature_table and md5sum check files. If the md5sum check fails, download the files again until it passes.
}
\examples{
## read in the dra (Deinococcus radiodurans R1) gff gz file in local disk
gzPath <- system.file("extdata", "dra.gff.gz", package = "ProGenome")
dragff <- read.gff(gzPath, isurl = FALSE, isgz = TRUE)

## only extract whole locus without genome division
locsRawMat <- ExtractLocs(dragff)

## whole locus divided by genomes and plasmids
locusList <- GetLocsfgff(dragff, genePrefix = 'dra:')

## get dra genomic locus through FTP URL
draLocs <- GetLocsfKEGGSpe('dra')

## two locus names for aac (Alicyclobacillus acidocaldarius subsp. acidocaldarius DSM 446)
gzPath <- system.file("extdata", "aac.gff.gz", package = "ProGenome")
aacgff <- read.gff(gzPath, isurl = FALSE, isgz = TRUE)
locusList <- GetLocsfgff(aacgff, genePrefix = 'aac:')

\dontrun{
hxaLocs <- GetLocsfKEGGSpe('hxa')
csuLocs <- GetLocsfKEGGSpe('csu')}
\dontrun{
download.SpeAnno('eco', 'gff', 'tmpEco')
}
}
\seealso{
read.gff read in gff files
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}
